safari-534.53-branch/LayoutTests/fast/js/string-substr-expected.txt

This test checks the boundary cases of substr().

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS 'bar'.substr(0) is 'bar'
PASS 'bar'.substr(3) is ''
PASS 'bar'.substr(4) is ''
PASS 'bar'.substr(-1) is 'r'
PASS 'bar'.substr(-3) is 'bar'
PASS 'bar'.substr(-4) is 'bar'
PASS 'bar'.substr(0, 0) is ''
PASS 'bar'.substr(0, 1) is 'b'
PASS 'bar'.substr(0, 3) is 'bar'
PASS 'bar'.substr(0, 4) is 'bar'
PASS 'bar'.substr(1, 0) is ''
PASS 'bar'.substr(1, 1) is 'a'
PASS 'bar'.substr(1, 2) is 'ar'
PASS 'bar'.substr(1, 3) is 'ar'
PASS 'bar'.substr(3, 0) is ''
PASS 'bar'.substr(3, 1) is ''
PASS 'bar'.substr(3, 3) is ''
PASS 'bar'.substr(4, 0) is ''
PASS 'bar'.substr(4, 1) is ''
PASS 'bar'.substr(4, 3) is ''
PASS 'bar'.substr(-1, 0) is ''
PASS 'bar'.substr(-1, 1) is 'r'
PASS 'bar'.substr(-3, 1) is 'b'
PASS 'bar'.substr(-3, 3) is 'bar'
PASS 'bar'.substr(-3, 4) is 'bar'
PASS 'bar'.substr(-4) is 'bar'
PASS 'bar'.substr(-4, 0) is ''
PASS 'bar'.substr(-4, 1) is 'b'
PASS 'bar'.substr(-4, 3) is 'bar'
PASS 'bar'.substr(-4, 4) is 'bar'
PASS 'GMAIL_IMP=bf-i%2Fd-0-0%2Ftl-v'.substr(10) is 'bf-i%2Fd-0-0%2Ftl-v'
PASS successfullyParsed is true

TEST COMPLETE
